requirements.txt
fastapi
google-generativeai
newspaper3k
uvicorn
pydantic
